requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
channels>=4.0.0
channels-redis>=4.1.0
Pillow>=10.0.0
ollama>=0.1.7
chromadb>=0.4.22
sentence-transformers>=2.5.0
requests>=2.31.0
numpy>=1.24.0
torch>=2.0.0


